+++ src/documentation-helpers.ts
@@ -17,8 +17,6 @@ import {
   isSchemaObject,
 } from "openapi3-ts/oas31";
 import {
-  both,
-  complement,
   concat,
   type as detectType,
   filter,
@@ -48,7 +46,6 @@ import {
   getExamples,
   hasCoercion,
   isCustomHeader,
-  isObject,
   makeCleanId,
   tryToTransform,
   ucFirst,
@@ -618,7 +615,7 @@ export const depictExamples = (
 ): ExamplesObject | undefined =>
   pipe(
     getExamples,
-    map(when(both(isObject, complement(Array.isArray)), omit(omitProps))),
+    map(when((subj) => detectType(subj) === "Object", omit(omitProps))),
     enumerateExamples,
   )({ schema, variant: isResponse ? "parsed" : "original", validate: true });