+++ src/logger-helpers.ts
@@ -25,3 +25,33 @@ export const severity: Record<keyof AbstractLogger, number> = {
 export const isLoggerInstance = (subject: unknown): subject is AbstractLogger =>
   isObject(subject) &&
   Object.keys(severity).some((method) => method in subject);
+
+/**
+ * @todo consider Intl units when Node 18 dropped (microsecond unit is missing, picosecond is not in list)
+ * @link https://tc39.es/ecma402/#table-sanctioned-single-unit-identifiers
+ * */
+const makeNumberFormat = (fraction = 0) =>
+  Intl.NumberFormat(undefined, {
+    useGrouping: false,
+    minimumFractionDigits: 0,
+    maximumFractionDigits: fraction,
+  });
+
+// creating them once increases the performance significantly
+const intFormat = makeNumberFormat();
+const floatFormat = makeNumberFormat(2);
+
+// not using R.cond for performance optimization
+const pickTimeUnit = (ms: number): [string, number, Intl.NumberFormat] => {
+  if (ms < 1e-6) return ["picosecond", ms / 1e-9, intFormat];
+  if (ms < 1e-3) return ["nanosecond", ms / 1e-6, intFormat];
+  if (ms < 1) return ["microsecond", ms / 1e-3, intFormat];
+  if (ms < 1e3) return ["millisecond", ms, intFormat];
+  if (ms < 6e4) return ["second", ms / 1e3, floatFormat];
+  return ["minute", ms / 6e4, floatFormat];
+};
+
+export const formatDuration = (durationMs: number) => {
+  const [unit, converted, formatter] = pickTimeUnit(durationMs);
+  return `${formatter.format(converted)} ${unit}${converted > 1 ? "s" : ""}`;
+};